6.1. Личный проект: исправил
